Add files via upload
This is the project I've been working on since the start of this week (1/20-1/24). This consists of a project library to understand how RDKit works, along with a mini protype of the final project. This initial prototype is helping me build proficiency with converting between user input and SMILES format to model images. So far, there is one class within this program, but it is not done yet. It can generate images of organic molecules, but it still lacks the proficiency to draw more complex molecules. I need to add the double and triple bond functionality.